# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: QuickApply Pro Admin Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /templates:
    get:
      operationId: getTemplates
      tags: [templates]
      summary: Get all email templates
      responses:
        "200":
          description: Templates list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplatesResponse"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create email template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateBody"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
  /templates/{templateId}:
    put:
      operationId: updateTemplate
      tags: [templates]
      summary: Update email template
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateBody"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete email template
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /templates/{templateId}/preview:
    post:
      operationId: previewTemplate
      tags: [templates]
      summary: Preview template with sample data
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rendered preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplatePreviewResponse"
  /templates/{templateId}/send-test:
    post:
      operationId: sendTestEmail
      tags: [templates]
      summary: Send test email to admin
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Test email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /campaigns/runs:
    get:
      operationId: getCampaignRuns
      tags: [campaigns]
      summary: Get campaign run history
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [inactive, newsletter]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Campaign runs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignRunsResponse"
...
  /campaigns/newsletter/run:
    post:
      operationId: runNewsletterCampaign
      tags: [campaigns]
      summary: Manually run newsletter campaign
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterRunBody"
      responses:
        "200":
          description: Campaign result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignRunResult"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
        success:
          type: boolean
      required: [message]
...
    EmailTemplate:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        campaignType:
          type: string
          enum: [inactive, newsletter]
        subject:
          type: string
        htmlBody:
          type: string
        textBody:
          type: string
        angle:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [_id, name, campaignType, subject, htmlBody]
    TemplateBody:
      type: object
      properties:
        name:
          type: string
        campaignType:
          type: string
          enum: [inactive, newsletter]
        subject:
          type: string
        htmlBody:
          type: string
        textBody:
          type: string
        angle:
          type: string
        isActive:
          type: boolean
      required: [name, campaignType, subject, htmlBody]
    TemplatesResponse:
      type: object
      properties:
        templates:
          type: array
          items:
            $ref: "#/components/schemas/EmailTemplate"
      required: [templates]
    TemplatePreviewResponse:
      type: object
      properties:
        subject:
          type: string
        htmlBody:
          type: string
        textBody:
          type: string
      required: [subject, htmlBody]
    CampaignRun:
      type: object
      properties:
        _id:
          type: string
        campaignType:
          type: string
          enum: [inactive, newsletter]
        triggerType:
          type: string
          enum: [manual, scheduled]
        startedAt:
          type: string
        finishedAt:
          type: string
        matchedUsers:
          type: integer
        sentCount:
          type: integer
        failedCount:
          type: integer
        skippedCount:
          type: integer
        status:
          type: string
          enum: [running, completed, failed]
        notes:
          type: string
      required: [_id, campaignType, triggerType, startedAt, sentCount, failedCount, skippedCount, status]
    CampaignRunsResponse:
      type: object
      properties:
        runs:
          type: array
          items:
            $ref: "#/components/schemas/CampaignRun"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [runs, total]
...
    CampaignRunResult:
      type: object
      properties:
        runId:
          type: string
        sentCount:
          type: integer
        failedCount:
          type: integer
        skippedCount:
          type: integer
        matchedUsers:
          type: integer
        status:
          type: string
        message:
          type: string
      required: [sentCount, failedCount, skippedCount, status]
    NewsletterRunBody:
      type: object
      properties:
        topic:
          type: string
        templateId:
          type: string
...